refactor(likes): rename files
